requirements.txt
langchain==0.2.6
langchain_anthropic==0.1.19
langchain_community==0.2.6
langchain_core==0.2.11
langchain_openai==0.1.14
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.30.0
